% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scanbase.R
\name{apply_xfm}
\alias{apply_xfm}
\title{Apply an xfm to a minc file}
\usage{
apply_xfm(xfm, input, output, like, clobber = TRUE)
}
\arguments{
\item{xfm}{The transform}

\item{input}{The minc to transform}

\item{output}{an output file}

\item{like}{a like file}
}
\value{
the transformed file invisibly
}
\description{
Apply an xfm to a minc file
}
