% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanbaseTools-package.R
\docType{package}
\name{scanbaseTools-package}
\alias{scanbaseTools}
\alias{scanbaseTools-package}
\title{scanbaseTools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
